+++ cinch_gen/CHANGELOG.md
@@ -1,5 +1,13 @@
 # cinch_gen changelog
 
+## 3.0.0-dev.2
+
+- Service Add params `sendTimeout` and `validateStatus`.
+
+## 3.0.0-dev
+
+- update dio to 5.0.0
+
 ## 2.2.0
 
 - update plugin